requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
psycopg2-binary>=2.9.9
flask>=3.0.0
flask-cors>=4.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
Pillow>=10.0.0
httpx>=0.27.0
# Poster detector (visión IA — Gemini Flash)
google-genai>=1.0.0
PyJWT>=2.8.0
gender-guesser>=0.4.0
